+++ backend/canisters/registry/impl/src/lib.rs
@@ -197,6 +197,7 @@ impl Data {
             ["ICRC-1".to_string(), "ICRC-2".to_string()].to_vec(),
             None,
             false,
+            Vec::new(),
             now,
         );
     }