+++ backend/external_canisters/one_sec_minter/api/src/queries/mod.rs
@@ -1,1 +1,2 @@
 pub mod get_forwarding_address;
+pub mod get_metadata;